requirements.txt
# Core dependencies - updated to compatible versions
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6

# Document processing
pypdf==4.2.0

# Vector store
faiss-cpu==1.8.0

# OpenAI
openai==1.26.0

# Utilities
python-dotenv==1.0.1
tiktoken==0.7.0